requirements.txt
python-binance==1.0.19
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
ta==0.10.2
google-api-python-client==2.108.0
newsapi-python==0.2.7
youtube-transcript-api==0.6.1
psutil==5.9.5
pytube==15.0.0
textblob==0.17.1
schedule==1.2.1
requests==2.31.0
ccxt==4.1.13
TA-Lib==0.6.4
aiohttp==3.9.1
asyncio==3.4.3 